feat(settings): move settings into the main workspace shell (#194)
Rename ChannelLayout to MainLayout and render the settings and teams
sections inside it. The dock swaps its channel list for a settings nav
(Profile, Security, Teams, Appearance, Notifications, Data & privacy)
so settings live in a single sidebar rather than a nested one, and the
content column owns its own scroll. Remove the now-orphaned starter-kit
app shell (AppLayout/AppSidebarLayout/AppSidebar, AppHeader, AppShell,
AppContent, NavMain/NavFooter, TeamSwitcher, Breadcrumbs, AppLogo) and
their dead AppVariant/BreadcrumbItem types.

Restructure every settings page into two-column sections (label left,
controls right) via a shared SettingsSection, wrap account deletion in a
Danger zone, and pull "Data & privacy" out of the Profile page into its
own settings item backed by a new data-export.edit route.
