+++ app/Http/Controllers/Settings/ProfileController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\Settings;
 
-use App\Data\DataExportData;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Settings\ProfileDeleteRequest;
 use App\Http\Requests\Settings\ProfileUpdateRequest;
@@ -21,12 +20,9 @@ class ProfileController extends Controller
      */
     public function edit(Request $request): Response
     {
-        $latestExport = $request->user()->dataExports()->first();
-
         return Inertia::render('settings/Profile', [
             'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
             'status' => $request->session()->get('status'),
-            'dataExport' => $latestExport === null ? null : DataExportData::fromExport($latestExport),
         ]);
     }
