+++ resources/js/layouts/settings/Layout.vue
@@ -1,79 +1,24 @@
-<script setup lang="ts">
-import { Link } from '@inertiajs/vue3';
-import { Separator } from '@/components/ui/separator';
-import { useCurrentUrl } from '@/composables/useCurrentUrl';
-import { toUrl } from '@/lib/utils';
-import { edit as editAppearance } from '@/routes/appearance';
-import { edit as editNotifications } from '@/routes/notifications';
-import { edit as editProfile } from '@/routes/profile';
-import { edit as editSecurity } from '@/routes/security';
-import { index as teams } from '@/routes/teams';
-import type { NavItem } from '@/types';
-
-const sidebarNavItems: NavItem[] = [
-    {
-        title: 'Profile',
-        href: editProfile(),
-    },
-    {
-        title: 'Security',
-        href: editSecurity(),
-    },
-    {
-        title: 'Teams',
-        href: teams(),
-    },
-    {
-        title: 'Appearance',
-        href: editAppearance(),
-    },
-    {
-        title: 'Notifications',
-        href: editNotifications(),
-    },
-];
-
-const { isCurrentOrParentUrl } = useCurrentUrl();
-</script>
-
 <template>
-    <div class="px-4 py-6">
-        <header class="mb-6 border-b border-border pb-5">
-            <h2
-                class="font-serif text-[32px] leading-none font-semibold tracking-tight"
-            >
-                Settings
-            </h2>
-            <p class="mt-1.5 text-sm text-muted-foreground">
-                Manage your profile and account settings
-            </p>
-        </header>
+    <!--
+      The settings navigation now lives in the workspace dock (see SettingsNav),
+      so this layout only frames the settings content: a scrollable column with a
+      section heading. Scrolling is owned here because the main card clips overflow.
+    -->
+    <div class="min-h-0 flex-1 overflow-y-auto px-6 py-8 lg:px-10">
+        <div class="w-full max-w-4xl">
+            <header class="mb-8 border-b border-border pb-5">
+                <h2
+                    class="font-serif text-[32px] leading-none font-semibold tracking-tight"
+                >
+                    Settings
+                </h2>
+                <p class="mt-1.5 text-sm text-muted-foreground">
+                    Manage your profile and account settings
+                </p>
+            </header>
 
-        <div class="flex flex-col lg:flex-row lg:space-x-12">
-            <aside class="w-full max-w-xl lg:w-48">
-                <nav class="flex flex-col gap-0.5" aria-label="Settings">
-                    <Link
-                        v-for="item in sidebarNavItems"
-                        :key="toUrl(item.href)"
-                        :href="item.href"
-                        :class="[
-                            'flex h-[34px] items-center rounded-lg px-3 text-sm font-medium transition-colors',
-                            isCurrentOrParentUrl(item.href)
-                                ? 'bg-primary text-primary-foreground'
-                                : 'text-muted-foreground hover:bg-accent hover:text-foreground',
-                        ]"
-                    >
-                        {{ item.title }}
-                    </Link>
-                </nav>
-            </aside>
-
-            <Separator class="my-6 lg:hidden" />
-
-            <div class="flex-1 md:max-w-2xl">
-                <section class="max-w-xl space-y-12">
-                    <slot />
-                </section>
+            <div class="space-y-10">
+                <slot />
             </div>
         </div>
     </div>
