+++ resources/js/types/navigation.ts
@@ -1,11 +1,6 @@
 import type { InertiaLinkProps } from '@inertiajs/vue3';
 import type { LucideIcon } from '@lucide/vue';
 
-export type BreadcrumbItem = {
-    title: string;
-    href: NonNullable<InertiaLinkProps['href']>;
-};
-
 export type NavItem = {
     title: string;
     href: NonNullable<InertiaLinkProps['href']>;
