+++ resources/js/types/ui.ts
@@ -1,8 +1,6 @@
 export type Appearance = 'light' | 'dark' | 'system';
 export type ResolvedAppearance = 'light' | 'dark';
 
-export type AppVariant = 'header' | 'sidebar';
-
 export type FlashToast = {
     type: 'success' | 'info' | 'warning' | 'error';
     message: string;
